repaired Win32 version

+++ src/Background.cpp
@@ -1,3 +1,5 @@
+#include "stdafx.h"
+
 #include "Background.h"
 
 #include "scaler.h"
@@ -5,23 +7,23 @@
 
 Background::Background (const char* filename, int width, int height)
 {
-  float stretch_x = g_Game.GetScreen_W() / (float)BASE_SCREEN_X;
-  float stretch_y = g_Game.GetScreen_H() / (float)BASE_SCREEN_Y;
+  float stretch_x = g_Game.GetScreen_W () / (float)BASE_SCREEN_X;
+  float stretch_y = g_Game.GetScreen_H () / (float)BASE_SCREEN_Y;
 
-  int new_width = width*stretch_x ;
-  int new_height = height*stretch_y;
+  int new_width  = static_cast<int>(width *stretch_x);
+  int new_height = static_cast<int>(height*stretch_y);
 
   image = LoadScaledBitmap (filename, new_width, new_height);
 
   Background::width = new_width;
   Background::height = new_height;
 
-  Background::x = g_GamePtr->GetScreen_W()/2.0;
-  Background::y = g_GamePtr->GetScreen_H()/2.0;
+  Background::x = static_cast<float>(g_GamePtr->GetScreen_W ()) / 2.0F;
+  Background::y = static_cast<float>(g_GamePtr->GetScreen_H ()) / 2.0F;
 
   if (image)
   {
-    clip = new SDL_Rect();
+    clip = new SDL_Rect ();
     clip->x = 0;
     clip->y = 0;
     clip->w = Background::width;
